#6365: Move comparison functions file into tt_dnn as production tt_eager code is importing from it and get rid of unnecessary models imports in tt_eager. Symlink test comparison funcs to tt_eager one that's new
