+++ tt_eager/tt_dnn/op_library/sliding_window_op_infra/sliding_window_op_config_generation_and_validation.py
@@ -4,7 +4,7 @@
 
 import torch
 import numpy
-from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_allclose_and_pcc
+from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal
 
 
 def generate_sliding_window_op_sharded_input_top_left_indices(
