+++ tt_eager/tt_dnn/op_library/sliding_window_op_infra/untilize_with_halo_config_generation_and_validation.py
@@ -6,7 +6,7 @@ import torch
 import numpy as np
 from loguru import logger
 
-from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_allclose_and_pcc
+from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal
 
 
 def trace_conv_to_generate_data_top_left_indices_and_pad_metadata(conv_params, input_nchw_shape):
